Implement tcc_no_lift, which tries to return the non-lifted effect of a term
When inspecting the effect of terms, tcc returns the potentially lifted
effects, which is not what we want because the original effectful information is
then cluttered. In the worst case, F* gives the effect information in terms of a
quite primitive effect, like STATE, which can't be easily exploited by
the user. Besides, it also causes problem for the functions which
explore the context to find suitable state variables with which to
instantiate pre/postconditions.
As this is mostly a problem for applications, the current workaround is,
in case of applications, to strip the parameters, retrieve the head's
effect with tcc, and instantiate this effect. In the other situations we
just use tcc.
Note that we tried inspecting then repacking the term just before the
call to tcc to check if it allows to "forget" about the context, but it
doesn't work.