Start improving speed by removing calls to trefl and as many calls to normalizer as possible
The calls to trefl are very expensive and useless. Instead, we admit the
proof (the system needs to be improved to make it more robust and track
errors).
Besides, the calls to the normalizer are very expensive. The problem is
that we use the normalizer a lot for variables substitutions. We thus
implemented our custom substitution functions.