Make the error detection more robust
In order to improve speed, we don't call trefl() at the end of the
post-processing anymore and abort on purpose. We thus need in elisp to
check whether F* failed on purpose or not: this is done by outputting
an indicator in the *Messages* buffer just before querying F*, and
outputting another indicator at the end of the F* output, just before
aborting: we then go to the end of the *Messages* buffer, look backward
for the first indicator, and look forward for the second one - its
presence indicates success, its absence indicates failure.