Change the way the type stored in typ_or_comp is instantiated
We can't use our custom apply_subst because it removes information used
by F* for pretty printing (for example, it turns `A /\ B` into
`Prims.l_and A B`) so we switched back to normalization. However, we
implemented a lazy mechanism to instantiate the variables in
typ_or_comp, with a `flush_typ_or_comp` function, so as to minimize the
calls to the normalizer.